Refactor: 1주차 dev 전체 병합 - User CRUD - Content CRUD - Comment CRUD

+++ src/main/java/com/java/everyboard/content/service/ContentService.java
@@ -1,45 +1,39 @@
 package com.java.everyboard.content.service;
 
+import com.java.everyboard.comment.repository.CommentRepository;
+import com.java.everyboard.content.dto.ContentAllResponseDto;
 import com.java.everyboard.content.dto.HomepageContentResponseDto;
 import com.java.everyboard.content.entity.Content;
 import com.java.everyboard.content.mapper.ContentMapper;
 import com.java.everyboard.content.repository.ContentRepository;
 import com.java.everyboard.exception.BusinessLogicException;
 import com.java.everyboard.exception.ExceptionCode;
+import com.java.everyboard.response.SingleResponseDto;
+import com.java.everyboard.user.entity.User;
+import com.java.everyboard.user.repository.UserRepository;
+import com.java.everyboard.user.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class ContentService {
-//    private final UserRepository userRepository;
-//    private final UserService userService;
+    private final UserRepository userRepository;
+    private final UserService userService;
     private final ContentRepository contentRepository;
-
     private final ContentMapper contentMapper;
-
-//    private final CommentRepository commentRepository;
-
-    public ContentService(ContentRepository contentRepository, ContentMapper contentMapper) {
-        this.contentRepository = contentRepository;
-        this.contentMapper = contentMapper;
-    }
-
-    /*public ContentService(UserRepository userRepository, UserService userService, ContentRepository contentRepository, ContentMapper contentMapper, CommentRepository commentRepository) {
-        this.userRepository = userRepository;
-        this.userService = userService;
-        this.contentRepository = contentRepository;
-        this.contentMapper = contentMapper;
-        this.commentRepository = commentRepository;
-    }*/
+    private final CommentRepository commentRepository;
 
     // 게시글 생성 //
     public Content createContent(Content content) {
-//        content.setUser(userService.getLoginMember());
-
+        content.setUser(userService.getLoginMember());
         return contentRepository.save(content);
     }
 
@@ -47,10 +41,10 @@ public class ContentService {
     public Content updateContent(Content content) {
         Content findContent = findVerifiedContent(content.getContentId());
 
-        /*User writer = userService.findVerifiedUser(findContent.getUser().getUserId()); // 작성자 찾기
+        User writer = userService.findVerifiedUser(findContent.getUser().getUserId()); // 작성자 찾기
         if(userService.getLoginMember().getUserId() != writer.getUserId()) // 작성자와 로그인한 사람이 다를 경우
             throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED); //예외 발생(권한 없음)
-*/
+
         Optional.ofNullable(content.getTitle())
                 .ifPresent(findContent::setTitle);
 
@@ -84,20 +78,20 @@ public class ContentService {
     public void deleteContent(Long contentId) {
         Content findContent = findVerifiedContent(contentId);
 
-//        User writer = userService.findVerifiedUser(findContent.getUser().getUserId()); // 작성자 찾기
-//        if(userService.getLoginMember().getUserId() != writer.getUserId()) // 작성자와 로그인한 사람이 다를 경우
-//            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED); //예외 발생(권한 없음)
+        User writer = userService.findVerifiedUser(findContent.getUser().getUserId()); // 작성자 찾기
+        if(userService.getLoginMember().getUserId() != writer.getUserId()) // 작성자와 로그인한 사람이 다를 경우
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED); //예외 발생(권한 없음)
         contentRepository.delete(findContent);
     }
 
-//    // 유저 검증 로직 //
-//    public User findVerifiedUser(Long userId) {
-//        Optional<User> optionalUser = userRepository.findById(userId);
-//        User findUser =
-//                optionalUser.orElseThrow(() ->
-//                        new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
-//        return findUser;
-//    }
+    // 유저 검증 로직 //
+    public User findVerifiedUser(Long userId) {
+        Optional<User> optionalUser = userRepository.findById(userId);
+        User findUser =
+                optionalUser.orElseThrow(() ->
+                        new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+        return findUser;
+    }
 
     // 게시글 검증 로직 //
     public Content findVerifiedContent(Long contentId) {
@@ -112,14 +106,14 @@ public class ContentService {
         return contentRepository.save(content);
     }
 
-//    @Transactional
-//    public ResponseEntity detail(Content content) {
-//
-//        ContentAllResponseDto response = contentMapper.contentToContentAllResponse(content, commentRepository);
-//
-//        return new ResponseEntity<>(
-//                new SingleResponseDto<>(response), HttpStatus.OK
-//        );
-//    }
+    @Transactional
+    public ResponseEntity detail(Content content) {
+
+        ContentAllResponseDto response = contentMapper.contentToContentAllResponse(content, commentRepository);
+
+        return new ResponseEntity<>(
+                new SingleResponseDto<>(response), HttpStatus.OK
+        );
+    }
 
 }
